parse-options: convert "command mode" to a flag
OPTION_CMDMODE is essentially OPTION_SET_INT plus the extra check
that the variable had not set before.  In order to allow custom
processing, change it to OPTION_SET_INT plus a new flag that takes
care of the check.  This works as long as the option value points
to an int.

Add testcases while at it.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>