am: support --show-current-patch=raw as a synonym for--show-current-patch
When "git am --show-current-patch" was added in commit 984913a210 ("am:
add --show-current-patch", 2018-02-12), "git am" started recommending it
as a replacement for .git/rebase-merge/patch.  Unfortunately the suggestion
is misguided, for example the output "git am --show-current-patch" cannot
be passed to "git apply" if it is encoded as quoted-printable or base64.

We would like therefore to add a new mode to "git am" that copies
.git/rebase-merge/patch to stdout.  In order to preserve backwards
compatibility, "git am --show-current-patch"'s behavior as to stay as
is, and the new functionality will be added as an optional
argument to --show-current-patch.  As a start, add the code to parse
submodes.  For now "raw" is the only valid submode, and it prints
the full e-mail message just like "git am --show-current-patch".

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>